+++ src/swift/core/interface/StatementSelectInterface.swift
@@ -23,7 +23,13 @@
  */
 
 import Foundation
+#if SWIFT_PACKAGE
 import WCDB_Private
+#elseif WCDB_SWIFT_TARGET
+import WCDBSwift.Private
+#else
+import WCDB.Private
+#endif
 
 /// Convenient interface for getting result with statement
 public protocol StatementSelectInterface: AnyObject {
@@ -67,6 +73,7 @@ extension StatementSelectInterface where Self: HandleRepresentable {
         let handle = try getHandle(writeHint: false)
         try handle.prepare(statement)
         guard try handle.step() else {
+            handle.finalize()
             return nil
         }
         let result = handle.value(atIndex: 0)
@@ -86,6 +93,7 @@ extension StatementSelectInterface where Self: HandleRepresentable {
         let handle = try getHandle(writeHint: false)
         try handle.prepare(statement)
         guard try handle.step() else {
+            handle.finalize()
             return nil
         }
         let result = handle.oneRowValue()